db, Header, App and Footer changes
